==> 0208TargetP2/1_merge_file.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!cp -r /content/drive/MyDrive/TargetP2_data /content/mywork

import os
os.chdir('/content/mywork')
current_path = os.getcwd()
os.listdir(current_path)

import pandas as pd

def fasta_to_dataframe(file_path):
    """将FASTA文件转换为pandas DataFrame"""
    # 初始化序列数据列表
    seq_data = []

    with open(file_path, 'r') as file:
        sequence_id = None
        sequence = ''
        for line in file:
            line = line.strip()
            if line.startswith(">"):
                if sequence_id is not None:
                    # 将当前序列添加到列表
                    seq_data.append([sequence_id, sequence])
                sequence_id = line[1:]  # 更新序列标识符
                sequence = ''  # 重置序列
            else:
                sequence += line  # 追加序列
        # 添加最后一个序列
        if sequence_id is not None:
            seq_data.append([sequence_id, sequence])

    # 创建DataFrame
    df = pd.DataFrame(seq_data, columns=['Entry', 'Sequence'])
    return df

file_path = 'targetp.fasta'
df1 = fasta_to_dataframe(file_path)
print(df1)



import numpy as np

def load_npz_file(file_path):
    # 加载npz文件
    data = np.load(file_path)

    # 从npz文件中读取数组
    fold = data['fold']
    ids = data['ids']
    y_type = data['y_type']

    # 定义y_type数字到文本标签的映射
    y_type_labels = {0: 'noTP', 1: 'SP', 2: 'mTP', 3: 'cTP', 4: 'luTP'}

    # 将y_type数组中的数字替换为对应的文本标签
    y_type_text = np.array([y_type_labels[y] for y in y_type])

    # 打印数组信息
    print("Fold:", fold)
    print("Entry:", ids)
    print("EC number:", y_type_text)

    # 关闭文件
    data.close()

    return fold, ids, y_type_text

# 调用函数并传入npz文件路径
file_path = 'targetp_data.npz'
folds, ids, y_type_text = load_npz_file(file_path)


# 将数组转换为Pandas DataFrame
data = {
    'Fold': folds,
    'Entry': ids,
    'EC Number': y_type_text
}
df2 = pd.DataFrame(data)
print(df2)

# 根据'Entry'列合并两个DataFrame
merged_df = pd.merge(df1, df2, on='Entry', how='inner')

# 显示合并后的DataFrame
print(merged_df)

merged_df.to_csv('result.csv', index=False)
