==> 0208TargetP2/2_partition.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!cp -r /content/drive/MyDrive/TargetP2_data /content/mywork

import os
os.chdir('/content/mywork')
current_path = os.getcwd()
os.listdir(current_path)

import pandas as pd
df = pd.read_csv('result.csv')
df

value_counts = df['EC Number'].value_counts()
print(value_counts)
# value_counts = df['Entry'].value_counts()
# print(value_counts)

def split_set(df, validation_partition=0, test_partition=1):
  validation_set = df.loc[df['Fold'] == validation_partition].drop(columns=['Fold'])
  test_set = df.loc[df['Fold'] == test_partition].drop(columns=['Fold'])
  train_set = df.loc[(df['Fold'] != test_partition) & (df['Fold'] != validation_partition)].drop(columns=['Fold'])
  # display(validation_set)
  # display(test_set)
  # display(train_set)
  return validation_set, test_set, train_set

def save_split(validation_set, test_set, train_set, validation_partition=0, test_partition=1):
  validation_set['Entry\tEC number\tSequence'] = validation_set['Entry'].astype(str) + '\t' + validation_set['EC Number'].astype(str) + '\t' + validation_set['Sequence'].astype(str)
  validation_set = validation_set[['Entry\tEC number\tSequence']]
  # display(validation_set)
  validation_set.to_csv('TP2_partition/TargetP2_validation_'+str(validation_partition)+'.csv', index=False)

  test_set['Entry\tEC number\tSequence'] = test_set['Entry'].astype(str) + '\t' + test_set['EC Number'].astype(str) + '\t' + test_set['Sequence'].astype(str)
  test_set = test_set[['Entry\tEC number\tSequence']]
  # display(test_set)
  test_set.to_csv('TP2_partition/TargetP2_test_'+str(test_partition)+'.csv', index=False)

  train_set['Entry\tEC number\tSequence'] = train_set['Entry'].astype(str) + '\t' + train_set['EC Number'].astype(str) + '\t' + train_set['Sequence'].astype(str)
  train_set = train_set[['Entry\tEC number\tSequence']]
  # display(test_set)
  train_set.to_csv('TP2_partition/TargetP2_train_'+str(validation_partition)+str(test_partition)+'.csv', index=False)

for i in range(5):
    valid_num = i
    if valid_num == 4:
        test_num = 0
    else:
        test_num = valid_num + 1
    validation_set, test_set, train_set = split_set(df, valid_num, test_num)
    save_split(validation_set, test_set, train_set, valid_num, test_num)

# print(df)
# _df = pd.read_csv('TP2_partition/test_0.csv')
# print(_df)
# _df = pd.read_csv('TP2_partition/validation_1.csv')
# print(_df)
# _df = pd.read_csv('TP2_partition/train_40.csv')
# print(_df)

alldf = df.drop(columns=['Fold'])
alldf

alldf['Entry\tEC number\tSequence'] = alldf['Entry'].astype(str) + '\t' + alldf['EC Number'].astype(str) + '\t' + alldf['Sequence'].astype(str)

alldf = alldf[['Entry\tEC number\tSequence']]

alldf.to_csv('TP2_partition/TargetP2.csv', index=False)

alldf

!cp -r /content/mywork/TP2_partition /content/drive/MyDrive/TargetP2_data
