==> 0208TargetP2/3_esm.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
os.chdir('/content/drive/MyDrive')
current_path = os.getcwd()
os.listdir(current_path)

!git clone https://github.com/Yichuan0712/CLEAN.git

import os
os.chdir('/content/drive/MyDrive/CLEAN/app')
current_path = os.getcwd()
os.listdir(current_path)

# import shutil

# # 替换以下路径为你想删除的文件夹路径
# folder_path = '/content/CLEAN/app/data/model'

# # 删除非空文件夹
# shutil.rmtree(folder_path)

# import shutil

# # 替换以下路径为你想删除的文件夹路径
# folder_path = '/content/CLEAN/app/data/esm_data'

# # 删除非空文件夹
# shutil.rmtree(folder_path)

# import os
# os.chdir('/content/CLEAN/app')
# current_path = os.getcwd()
# os.listdir(current_path)

!pip install fair-esm  # latest release
!git clone https://github.com/facebookresearch/esm.git

from src.CLEAN.utils import csv_to_fasta, mutate_single_seq_ECs, retrive_esm1b_embedding, compute_esm_distance
csv_to_fasta("data/TargetP2.csv", "data/TargetP2.fasta")

!python esm/scripts/extract.py esm2_t6_8M_UR50D data/TargetP2.fasta data/esm_data --include mean

# !cp -r /content/CLEAN/app/data/esm_data /content/drive/MyDrive/esm_TargetP2

# 压缩名为folder_name的文件夹到archive_name.zip
# !zip -r /content/drive/MyDrive/esm_data.zip /content/drive/MyDrive/esm_data
import shutil
shutil.make_archive('/content/drive/MyDrive/CLEAN/app/data/test', 'zip', '/content/drive/MyDrive/CLEAN/app/data/esm_data')

https://stackoverflow.com/questions/54660967/google-colab-having-problems-with-drive-folders-containing-lots-of-files

Shit happens when you have too many files to move to Google Drive
