==> 0208TargetP2/5_test.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!git clone https://github.com/Yichuan0712/CLEAN.git

import os
os.chdir('/content/CLEAN/app')
current_path = os.getcwd()
os.listdir(current_path)

!pip install fair-esm  # latest release

from src.CLEAN.utils import csv_to_fasta

# 创建一个列表，包含所有要转换的文件名前缀
file_prefixes = ["TargetP2_test_", "TargetP2_validation_"]

# 对每个前缀进行循环
for prefix in file_prefixes:
    # 对于每种类型（测试和验证），都有5个文件，编号从0到4
    for i in range(5):
        # 构建csv文件名和fasta文件名
        csv_file = f"data/{prefix}{i}.csv"
        fasta_file = f"data/{prefix}{i}.fasta"
        # 转换文件
        csv_to_fasta(csv_file, fasta_file)

dict_list = []
from src.CLEAN.infer import infer_maxsep_triplet
for i in range(5):
  vid = str(i)
  if i!=4:
    tid = str(i+1)
  else:
    tid = '0'
  print()
  print('validation_'+vid,"test_"+tid)
  print()
  train_data = 'TargetP2_train_' + vid + tid
  test_data = "TargetP2_validation_" + vid
  m_name = 'TargetP2_triplet_'+vid+tid

  dict1 = {'Loss': 'triplet', 'Selection': 'maxsep', 'Fold':'validation_'+vid}
  dict11 = infer_maxsep_triplet(train_data, test_data, report_metrics=True, pretrained=False, model_name=m_name)
  dict1.update(dict11)

  test_data = "TargetP2_test_" + tid
  dict2 = {'Loss': 'triplet', 'Selection': 'maxsep', 'Fold':"test_"+tid}
  dict22 = infer_maxsep_triplet(train_data, test_data, report_metrics=True, pretrained=False, model_name=m_name)
  dict2.update(dict22)

  dict_list.append(dict1)
  dict_list.append(dict2)
dict_list

from src.CLEAN.infer import infer_pvalue_triplet
for i in range(5):
  vid = str(i)
  if i!=4:
    tid = str(i+1)
  else:
    tid = '0'
  print()
  print('validation_'+vid,"test_"+tid)
  print()
  train_data = 'TargetP2_train_' + vid + tid
  test_data = "TargetP2_validation_" + vid
  m_name = 'TargetP2_triplet_'+vid+tid

  dict1 = {'Loss': 'triplet', 'Selection': 'pvalue', 'Fold':'validation_'+vid}
  dict11 = infer_pvalue_triplet(train_data, test_data, p_value=1e-5, nk_random=20, report_metrics=True, pretrained=False, model_name=m_name)
  dict1.update(dict11)

  test_data = "TargetP2_test_" + tid
  dict2 = {'Loss': 'triplet', 'Selection': 'pvalue', 'Fold':"test_"+tid}
  dict22 = infer_pvalue_triplet(train_data, test_data, p_value=1e-5, nk_random=20, report_metrics=True, pretrained=False, model_name=m_name)
  dict2.update(dict22)

  dict_list.append(dict1)
  dict_list.append(dict2)
dict_list

from src.CLEAN.infer import infer_maxsep_supcon
for i in range(5):
  vid = str(i)
  if i!=4:
    tid = str(i+1)
  else:
    tid = '0'
  print()
  print('validation_'+vid,"test_"+tid)
  print()
  train_data = 'TargetP2_train_' + vid + tid
  test_data = "TargetP2_validation_" + vid
  m_name = 'TargetP2_supcon_'+vid+tid

  dict1 = {'Loss': 'supcon', 'Selection': 'maxsep', 'Fold':'validation_'+vid}
  dict11 = infer_maxsep_supcon(train_data, test_data, report_metrics=True, pretrained=False, model_name=m_name)
  dict1.update(dict11)

  test_data = "TargetP2_test_" + tid
  dict2 = {'Loss': 'supcon', 'Selection': 'maxsep', 'Fold':"test_" + tid}
  dict22 = infer_maxsep_supcon(train_data, test_data, report_metrics=True, pretrained=False, model_name=m_name)
  dict2.update(dict22)

  dict_list.append(dict1)
  dict_list.append(dict2)
dict_list

from src.CLEAN.infer import infer_pvalue_supcon
for i in range(5):
  vid = str(i)
  if i!=4:
    tid = str(i+1)
  else:
    tid = '0'
  print()
  print('validation_'+vid,"test_"+tid)
  print()
  train_data = 'TargetP2_train_' + vid + tid
  test_data = "TargetP2_validation_" + vid
  m_name = 'TargetP2_supcon_'+vid+tid

  dict1 = {'Loss': 'supcon', 'Selection': 'pvalue', 'Fold':'validation_'+vid}
  dict11 = infer_pvalue_supcon(train_data, test_data, p_value=1e-5, nk_random=20, report_metrics=True, pretrained=False, model_name=m_name)
  dict1.update(dict11)
  test_data = "TargetP2_test_" + tid

  dict2 = {'Loss': 'supcon', 'Selection': 'pvalue', 'Fold':"test_" + tid}
  dict22 = infer_pvalue_supcon(train_data, test_data, p_value=1e-5, nk_random=20, report_metrics=True, pretrained=False, model_name=m_name)
  dict2.update(dict22)
  dict_list.append(dict1)
  dict_list.append(dict2)
dict_list

import pandas as pd
pd.DataFrame(dict_list).to_csv('output.csv', index=False)
